Move the temporary packages to thirdparty
License: MIT
